requirements.txt
Flask==2.3.2
requests==2.31.0
selenium==4.15.2
tqdm==4.66.1
webdriver-manager==4.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
google-generativeai==0.3.2
supabase==2.3.0
Pillow==10.2.0
openai==1.12.0